E_ALL - minor corrections to isset usage

+++ modules/Products/Products.php
@@ -88,18 +88,18 @@ class Products extends CRMEntity {
 	function save_module($module)
 	{
 		//Inserting into product_taxrel table
-		if(isset($_REQUEST['ajxaction']) != 'DETAILVIEW' && isset($_REQUEST['action']) != 'ProcessDuplicates' && !$this->isWorkFlowFieldUpdate)
+		if((isset($_REQUEST['ajxaction']) && $_REQUEST['ajxaction'] != 'DETAILVIEW') && (isset($_REQUEST['action']) && $_REQUEST['action'] != 'ProcessDuplicates') && !$this->isWorkFlowFieldUpdate)
 		{
-			if (isset($_REQUEST['ajxaction']) != 'CurrencyUpdate') {
+			if (isset($_REQUEST['ajxaction']) && $_REQUEST['ajxaction'] != 'CurrencyUpdate') {
 				$this->insertTaxInformation('vtiger_producttaxrel', 'Products');
 			}
 
-			if (isset($_REQUEST['action']) != 'MassEditSave' ) {
+			if (isset($_REQUEST['action']) && $_REQUEST['action'] != 'MassEditSave' ) {
 				$this->insertPriceInformation('vtiger_productcurrencyrel', 'Products');
 			}
 		}
 
-		if(isset($_REQUEST['action']) == 'SaveAjax' && isset($_REQUEST['base_currency']) && isset($_REQUEST['unit_price'])){
+		if((isset($_REQUEST['action']) && $_REQUEST['action'] == 'SaveAjax') && isset($_REQUEST['base_currency']) && isset($_REQUEST['unit_price'])){
 			$this->insertPriceInformation('vtiger_productcurrencyrel', 'Products');
 		}
 		// Update unit price value in vtiger_productcurrencyrel
@@ -123,7 +123,7 @@ class Products extends CRMEntity {
 		$tax_per = '';
 		//Save the Product - tax relationship if corresponding tax check box is enabled
 		//Delete the existing tax if any
-		if($this->mode == 'edit' && isset($_REQUEST['action']) != 'MassEditSave')
+		if($this->mode == 'edit' && (isset($_REQUEST['action']) && $_REQUEST['action'] != 'MassEditSave'))
 		{
 			for($i=0;$i<php7_count($tax_details);$i++)
 			{
@@ -141,8 +141,8 @@ class Products extends CRMEntity {
 				$taxid = getTaxId($tax_name);
 				$tax_per = $_REQUEST[$tax_name];
 
-				$taxRegions = isset($_REQUEST[$tax_name.'_regions']);
-				if ($taxRegions || isset($_REQUEST[$tax_name.'_defaultPercentage']) != '') {
+				$taxRegions = isset($_REQUEST[$tax_name.'_regions']) ? $_REQUEST[$tax_name.'_regions']:"";
+				if ($taxRegions || (isset($_REQUEST[$tax_name.'_defaultPercentage']) && $_REQUEST[$tax_name.'_defaultPercentage'] != '')) {
 					$tax_per = $_REQUEST[$tax_name.'_defaultPercentage'];
 				} else {
 					$taxRegions = array();
@@ -156,7 +156,7 @@ class Products extends CRMEntity {
 
 				$log->debug("Going to save the Product - $tax_name tax relationship");
 
-				if (isset($_REQUEST['action']) === 'MassEditSave') {
+				if (isset($_REQUEST['action']) && $_REQUEST['action'] === 'MassEditSave') {
 					$adb->pquery('DELETE FROM vtiger_producttaxrel WHERE productid=? AND taxid=?', array($this->id, $taxid));
 				}
 
@@ -182,7 +182,7 @@ class Products extends CRMEntity {
 		$currency_details = getAllCurrencies('all');
 
 		//Delete the existing currency relationship if any
-		if($this->mode == 'edit' && isset($_REQUEST['action']) !== 'CurrencyUpdate')
+		if($this->mode == 'edit' && (isset($_REQUEST['action']) && $_REQUEST['action'] !== 'CurrencyUpdate'))
 		{
 			for($i=0;$i<php7_count($currency_details);$i++)
 			{
@@ -205,7 +205,7 @@ class Products extends CRMEntity {
 			$requestPrice = CurrencyField::convertToDBFormat($_REQUEST['unit_price'], null, true);
 			$actualPrice = CurrencyField::convertToDBFormat($_REQUEST[$cur_valuename], null, true);
 			$isQuickCreate = false;
-			if(isset($_REQUEST['action'])=='SaveAjax' && isset($_REQUEST['base_currency']) && $_REQUEST['base_currency'] == $cur_valuename){
+			if((isset($_REQUEST['action']) && $_REQUEST['action'] =='SaveAjax') && isset($_REQUEST['base_currency']) && $_REQUEST['base_currency'] == $cur_valuename){
 				$actualPrice = $requestPrice;
 				$isQuickCreate = true;
 			}
@@ -218,7 +218,7 @@ class Products extends CRMEntity {
 
 				$log->debug("Going to save the Product - $curname currency relationship");
 
-				if (isset($_REQUEST['action']) === 'CurrencyUpdate') {
+				if (isset($_REQUEST['action']) && $_REQUEST['action'] === 'CurrencyUpdate') {
 					$adb->pquery('DELETE FROM vtiger_productcurrencyrel WHERE productid=? AND currencyid=?', array($this->id, $curid));
 				}
